agriculture-rag/requirements.txt
streamlit
langchain
langchain-community
pypdf
pdf2image
pillow
python-dotenv
openai
sentence-transformers
faiss-cpu
transformers
torch
tiktoken
pymupdf
numpy
tqdm